requirements.txt
exa-py
openai
faiss-cpu
numpy
python-dotenv
langsmith
uuid